ERP-9: Refactor code
