+++ COMP1640/Controllers/IdeaController.cs
@@ -1,4 +1,5 @@
 ﻿using COMP1640.Services;
+using COMP1640.ViewModels.Category.Requests;
 using COMP1640.ViewModels.Idea.Requests;
 using COMP1640.ViewModels.Idea.Responses;
 using Microsoft.AspNetCore.Http;
@@ -7,13 +8,15 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace COMP1640.Controllers
 {
+    [Route("idea")]
     public class IdeaController : Controller
     {
         private readonly IdeaService _ideaService;
-
-        public IdeaController(IdeaService ideaService)
+        private readonly CategoryService _categoryService;
+        public IdeaController(IdeaService ideaService, CategoryService categoryService)
         {
             _ideaService = ideaService;
+            _categoryService = categoryService;
         }
         [HttpGet]
         public async Task<IActionResult> Create()
@@ -37,6 +40,22 @@ namespace COMP1640.Controllers
 
             ModelState.AddModelError("create_exception", "Failure to create a new idea.");
             return View();
+        }
+        
+        [HttpGet]
+        [Route("viewcategory")]
+        public async Task<IActionResult> ViewCategory([FromQuery] GetListCategoryRequest request)
+        {
+            var category = await _categoryService.GetListCategory(request);
+            return View(category);
+        }
+
+        [HttpPost]
+        [Route("createcategory")]
+        public async Task<IActionResult> CreateCategory(CreateCategoryRequest request)
+        {
+            await _categoryService.CreateCategory(request);
+            return RedirectToAction("ViewCategory");
         }
     }
 }
