+++ COMP1640/ViewModels/Category/Requests/CreateCategoryRequest.cs
@@ -2,7 +2,5 @@ namespace COMP1640.ViewModels.Category.Requests;
 
 public class CreateCategoryRequest
 {
-    public int Id { get; set; }
     public string Name { get; set; }
-    public int TenantId { get; set; }
 }
