+++ COMP1640/Views/Idea/ViewCategory.cshtml
@@ -44,11 +44,6 @@
                                    required="required" />
                         </div>
                         <div class="mb-3">
-                            <input class="form-control"
-                                   type="text" name="tenantId" placeholder="Tenant Id"
-                                   required="required" />
-                            </div>
-                        <div class="mb-3">
                             <button class="btn btn-primary shadow d-block w-100" type="submit">Add</button>
                         </div>
                     </form>
