+++ Domain/Category.cs
@@ -9,6 +9,11 @@ namespace Domain
 
         }
 
+        public Category(string name)
+        {
+            Name = name;
+        }
+
         public string Name { get; set; }
 
         public virtual ICollection<Idea> Ideas { get; set; } = new HashSet<Idea>();
