+++ Domain/Interfaces/IBaseRepository.cs
@@ -8,10 +8,10 @@ namespace Domain.Interfaces
         IQueryable<T> GetQuery(Expression<Func<T, bool>> expression);
         Task<bool> AnyAsync(Expression<Func<T, bool>> expression);
         Task<T> GetAsync(Expression<Func<T, bool>> expression);
-        Task InsertAsync(T entity, bool saveChanges = true);
+        Task InsertAsync(T entity);
 
-        Task DeleteAsync(T entity, bool saveChanges = true);
+        Task DeleteAsync(T entity);
 
-        Task DeleteRangeAsync(IEnumerable<T> entities, bool saveChanges = true);
+        Task DeleteRangeAsync(IEnumerable<T> entities);
     }
 }
