+++ Domain/Interfaces/ICategoryRepository.cs
@@ -3,6 +3,6 @@
     public interface ICategoryRepository : IBaseRepository<Category>
     {
         Task<Category> GetAsync(int categoryId);
-        Task<Category> Add(Category category);
+        
     }
 }
